requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
anyio==4.0.0
    # via httpcore
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
exceptiongroup==1.1.3
    # via anyio
greenlet==3.0.0
    # via sqlalchemy
h11==0.14.0
    # via httpcore
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via python-telegram-bot
idna==3.4
    # via
    #   anyio
    #   httpx
python-dotenv==1.0.0
    # via -r requirements.in
python-telegram-bot==20.6
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.21
    # via -r requirements.in
typing-extensions==4.8.0
    # via sqlalchemy
